output/kms: Allow setting the variable refresh rate mode for an output
Add a function that enables variable refresh rate on a CRTC when it
is enabled and requested for the given output. If it isn't requested
or enabled for the output, it is disabled on the CRTC.

An output is considered capable of variable refresh rate if its
connector supports it.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1154