wayland/text-input: Open the OSK after setting input focus
Otherwise the OSK remains closed when only a single
zwp_text_input_v3.enable request for a given surface is received from
a client.

The previous behavior required focusing (or pressing) text input boxes
twice for the OSK to open.

This undoes a part of eabb789381f6 which doesn't appear to be
necessary to accomplish its purpose of keeping the OSK open on input
focus change within a surface.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1506

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1543